incorporate feedback - remove custom stream assembly for subscriptions
